Fix rare possible hang ups (nextMoveAllowed issue)
In rare cases a client "hangs up" waiting for Ack to send next Move
command in case of connection problems (see v. 1.1.1);
clicking on Swagga button fixes the situation, but only this action!
Because a user cannot click on Move as far as it is blocked;
So since 1.1.6 it will be unblocked any way within 10 sec.